+++ src/command/commands/chat.ts
@@ -18,7 +18,7 @@ const command: Command = {
     }
 
     // Make any necessary changes based on the platform
-    if (ctx.platform === 'twitch') {
+    if (ctx.platform === Platform.Twitch) {
       // Doesn't respond with the owners name if the owner is the one who sent the message
       // This helps with ownerSendCommandDirectly for things like !ask so it doesn't look like
       // blunderBot is responding to itself or the owner